Minor changes to pass a few datamodels

+++ python/datamodel_parser/application/Store.py
@@ -154,11 +154,13 @@ class Store():
         self.path_info = None
         if self.ready:
             if self.filepath:
-                # set datamodel path, self.env_variable, self.file_name, and self.directory_names
+                # set datamodel path, self.env_variable, self.file_name,
+                # and self.directory_names
                 path = self.filepath
                 self.set_path(path=path)
                 self.split_path()
-                # self.populate_file_path_tables() # this is done in populate_filespec_table_archive
+                # self.populate_file_path_tables() # this is done in the method
+                # populate_filespec_table_archive
                 if self.ready:
                     self.path_info = {'filepath'         : self.filepath        ,
                                       'env_variable'     : self.env_variable    ,
@@ -261,20 +263,20 @@ class Store():
                     yaml_file = (join(self.yaml_dir,filename)
                                  if self.yaml_dir else None)
 
-                    # DEBUG #
-                    with open(yaml_file,'w') as file:
-                        file.write(yaml_str)
-                    # DEBUG #
+#                    # DEBUG #
+#                    with open(yaml_file,'w') as file:
+#                        file.write(yaml_str)
+#                    # DEBUG #
 
-#                    if not exists(yaml_file):
-#                        with open(yaml_file,'w') as file:
-#                            file.write(yaml_str)
-#                    else:
-#                        self.ready = False
-#                        self.logger.error(
-#                            'Unable to init_location_substitutions_yaml. ' +
-#                            'The file already exists: {}. '.format(yaml_file) +
-#                            'If you wish to replace this file please first delete it.')
+                    if not exists(yaml_file):
+                        with open(yaml_file,'w') as file:
+                            file.write(yaml_str)
+                    else:
+                        self.ready = False
+                        self.logger.error(
+                            'Unable to init_location_substitutions_yaml. ' +
+                            'The file already exists: {}. '.format(yaml_file) +
+                            'If you wish to replace this file please first delete it.')
             else:
                 self.ready = False
                 self.logger.error('Unable to init_location_substitutions_yaml. ' +
